refactor out snapshot code
