+++ src/main/java/bio/terra/service/snapshot/flight/create/CountSnapshotTableRowsStep.java
@@ -3,7 +3,7 @@ package bio.terra.service.snapshot.flight.create;
 import bio.terra.model.SnapshotRequestModel;
 import bio.terra.service.snapshot.Snapshot;
 import bio.terra.service.snapshot.SnapshotDao;
-import bio.terra.service.tabulardata.google.bigquery.BigQueryPdao;
+import bio.terra.service.tabulardata.google.bigquery.BigQuerySnapshotPdao;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.Step;
 import bio.terra.stairway.StepResult;
@@ -12,13 +12,15 @@ import java.util.Map;
 
 public class CountSnapshotTableRowsStep implements Step {
 
-  private final BigQueryPdao bigQueryPdao;
+  private final BigQuerySnapshotPdao bigQuerySnapshotPdao;
   private final SnapshotDao snapshotDao;
   private final SnapshotRequestModel snapshotReq;
 
   public CountSnapshotTableRowsStep(
-      BigQueryPdao bigQueryPdao, SnapshotDao snapshotDao, SnapshotRequestModel snapshotReq) {
-    this.bigQueryPdao = bigQueryPdao;
+      BigQuerySnapshotPdao bigQuerySnapshotPdao,
+      SnapshotDao snapshotDao,
+      SnapshotRequestModel snapshotReq) {
+    this.bigQuerySnapshotPdao = bigQuerySnapshotPdao;
     this.snapshotDao = snapshotDao;
     this.snapshotReq = snapshotReq;
   }
@@ -27,7 +29,7 @@ public class CountSnapshotTableRowsStep implements Step {
   public StepResult doStep(FlightContext flightContext)
       throws InterruptedException, RetryException {
     Snapshot snapshot = snapshotDao.retrieveSnapshotByName(snapshotReq.getName());
-    Map<String, Long> tableRowCounts = bigQueryPdao.getSnapshotTableRowCounts(snapshot);
+    Map<String, Long> tableRowCounts = bigQuerySnapshotPdao.getSnapshotTableRowCounts(snapshot);
     snapshotDao.updateSnapshotTableRowCounts(snapshot, tableRowCounts);
     return StepResult.getStepResultSuccess();
   }
